requirements.txt
hmmlearn>=0.3.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
scipy>=1.12.0
alpaca-py>=0.20.0
python-dotenv>=1.0.0
streamlit>=1.33.0
plotly>=5.20.0
yfinance>=0.2.38
ta>=0.11.0
requests>=2.31.0
python-dateutil>=2.9.0
pytz>=2024.1
